use 64 bit fixed point for stepticker

+++ src/modules/robot/Block.h
@@ -51,11 +51,11 @@ class Block {
 
         // this is the data needed to determine when each motor needs to be issued a step
         using tickinfo_t= struct {
-            int32_t steps_per_tick; // 2.30 fixed point
-            int32_t counter; // 2.30 fixed point
-            int32_t acceleration_change; // 2.30 fixed point signed
-            int32_t deceleration_change; // 2.30 fixed point
-            int32_t plateau_rate; // 2.30 fixed point
+            int64_t steps_per_tick; // 2.62 fixed point
+            int64_t counter; // 2.62 fixed point
+            int64_t acceleration_change; // 2.62 fixed point signed
+            int64_t deceleration_change; // 2.62 fixed point
+            int64_t plateau_rate; // 2.62 fixed point
             uint32_t steps_to_move;
             uint32_t step_count;
             uint32_t next_accel_event;
